separate out working code from init
